ia3d:: A3DCreator: load A3DMMesh

+++ ia3d/src/main/java/good/damn/ia3d/stream/A3DInputStream.java
@@ -28,7 +28,7 @@ extends DataInputStream {
         mBuffer = new byte[16];
     }
 
-    public short readLUShort() throws IOException {
+    public int readLUShort() throws IOException {
         read(
             mBuffer,
             0, 2
