+++ ia3d/src/main/java/good/damn/ia3d/utils/A3DUtils.java
@@ -10,14 +10,12 @@ public final class A3DUtils {
 
     private static final Charset UTF_8 = Charset.forName("UTF-8");
 
-    public static short readUnsignedShortL(
+    public static int readUnsignedShortL(
         byte[] buffer,
         int offset
     ) {
-        return (short) (
-            buffer[offset+1] & 0xff << 8 |
-            buffer[offset] & 0xff
-        );
+        return buffer[offset+1] & 0xff << 8 |
+            buffer[offset] & 0xff;
     }
 
     public static String readNullTerminatedString(
